Simplify building tests
Instead of building a bunch of do_check executables, build a single
genericlisp_tests executable to run all of the tests, and use suffix rules to
ensure the tests are built appropriately.

Also use pkg-config (assuming it’s in the user’s PATH) to generate the CFLAGS
and LDFLAGS used to link check.