Xcode: Replace XCTest with check (not run)
Since the tests were ported to check, there’s no utility in keeping around the
file references to the old tests that were replaced, so swap in the check
tests.

Unfortunately it’s not straightforward to tell XCTest about the check tests in
a way that allows them to be run and for results to be reported through Xcode.
But at least they can be built--and run through the static analyzer.